requirements.txt
anyio==3.5.0
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cookiecutter==1.7.3
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.1
fastapi-crudrouter==0.8.5
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jinja2-time==0.2.0
manage-fastapi==1.1.1
MarkupSafe==2.1.1
orjson==3.6.7
poyo==0.5.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
python-slugify==6.1.1
PyYAML==5.4.1
questionary==1.10.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
sqlalchemy2-stubs==0.0.2a21
sqlmodel==0.0.6
starlette==0.17.1
text-unidecode==1.3
typer==0.4.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.9
uvicorn==0.15.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.2
